Better explaination of APP_DIR support
The engine already supported the idea of loading typst templates from
within Django app dirs, but it is was not clear how. The standard Django
template engine assumes if a app is call `example` then templates can be
found in `example/templates`.

So this is what people expect with this engine too. However, Django's
support for multiple engines has each engine expect a different folder
name within the app. For example the Jinja2 engine expected `example/jinja2`

We expect `example/typst`. So I have added a test to demonstrate this,
as well as added a note to the README,nd to make clearer.

Thanks to @ticosax for both raising this issue and providing the
foundations on which this could be built.

Fixes #22